api/utils/jwt_tokens.py
"""JWT access tokens."""

from datetime import datetime, timedelta, timezone

from jose import JWTError, jwt

from api.config import settings


def create_access_token(subject_user_id: int) -> str:
    expire = datetime.now(timezone.utc) + timedelta(minutes=settings.JWT_ACCESS_EXPIRE_MINUTES)
    to_encode = {
        "sub": str(subject_user_id),
        "exp": int(expire.timestamp()),
    }
    return jwt.encode(
        to_encode,
        settings.JWT_SECRET_KEY,
        algorithm=settings.JWT_ALGORITHM,
    )


def decode_access_token(token: str) -> int:
    """
    Validate JWT and return user_id from subject claim.
    Raises jose.JWTError on failure.
    """
    payload = jwt.decode(
        token,
        settings.JWT_SECRET_KEY,
        algorithms=[settings.JWT_ALGORITHM],
    )
    sub = payload.get("sub")
    if sub is None:
        raise JWTError("Missing subject")
    return int(sub)
