requirements.txt
# Core Dependencies
fastapi
uvicorn
pydantic
pydantic-settings
email-validator>=2.0.0

# Database
mysql-connector-python
pymongo==4.6.0

# Kafka
kafka-python==2.0.2

# Spark
pyspark==3.5.0

# Machine Learning
transformers
torch
scikit-learn

# Data Processing
pandas
numpy
pyarrow>=14.0.0

# Utilities
python-dotenv==1.0.0

# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

